+++ api/python/provisioner/commands/bootstrap/node.py
